Enhance CNN model for 90%+ accuracy with transfer learning
